visualizer/requirements.txt
# Dépendances Python - Programme de visualisation PC

# Communication série
pyserial==3.5

# Rendu graphique (choisir l'un)
pygame-ce>=2.5.6       # Graphiques 2D (Community Edition, compatible Python 3.14+)
# moderngl==5.10.0     # OpenGL (haute performance, effets spectaculaires)
# PyOpenGL==3.1.7      # OpenGL (solution traditionnelle)

# Calcul numérique
numpy==1.26.4

# Optionnel : surveillance des performances
# psutil==5.9.8
